azure-billing-archival/functions/integrity_check_function/requirements.txt

azure-functions
azure-storage-blob
